+++ src/init.c
@@ -60,6 +60,7 @@ static const R_CallMethodDef call_methods[] = {
   {"C_zuh_node_strings", (DL_FUNC) &C_zuh_node_strings, 2},
   {"C_zuh_select", (DL_FUNC) &C_zuh_select, 5},
   {"C_zuh_text_clean", (DL_FUNC) &C_zuh_text_clean, 3},
+  {"C_zuh_markdown", (DL_FUNC) &C_zuh_markdown, 4},
   {"C_zuh_outermost", (DL_FUNC) &C_zuh_outermost, 2},
   {"C_zuh_table", (DL_FUNC) &C_zuh_table, 3},
   {NULL, NULL, 0}